refactor: flatcoin constants
