+++ libs/flatcoin/constants/README.md
@@ -1,0 +1,7 @@
+# flatcoin-constants
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test flatcoin-constants` to execute the unit tests via [Vitest](https://vitest.dev/).
