+++ libs/flatcoin/constants/src/abis/FlatcoinOracleModuleABI.ts
