+++ libs/flatcoin/constants/src/abis/FlatcoinStableModuleABI.ts
