+++ libs/flatcoin/constants/src/config.ts
