+++ libs/flatcoin/constants/src/contracts/index.ts
@@ -1,0 +1,1 @@
+export * from './baseGoerli';
