+++ libs/flatcoin/constants/src/tokens/index.ts
@@ -1,0 +1,1 @@
+export * from './baseGoerli';
