+++ libs/flatcoin/constants/src/trading.ts
