+++ libs/flatcoin/page/src/components/Positions/Leverage/CloseLeveragePositionButton.tsx
@@ -1,4 +1,4 @@
-import { DEFAULT_MAX_FILL_PRICE_SLIPPAGE } from '@frontend/shared-constants';
+import { DEFAULT_MAX_FILL_PRICE_SLIPPAGE } from '@frontend/flatcoin-constants';
 import { usePushNotification } from '@frontend/shared-providers';
 import { TransactionActionButton } from '@frontend/shared-ui';
 import BigNumber from 'bignumber.js';
