+++ libs/flatcoin/page/src/components/Positions/Leverage/LeveragePositionsTable.tsx
@@ -1,5 +1,5 @@
 import { formatToUsd } from '@dhedge/core-ui-kit/utils';
-import { DEFAULT_TOKEN_DECIMALS } from '@frontend/shared-constants';
+import { DEFAULT_TOKEN_DECIMALS } from '@frontend/flatcoin-constants';
 import { TokenIconRevamp } from '@frontend/shared-ui';
 import {
   formatNumberToLimitedDecimals,
