+++ libs/shared/constants/src/abis/index.ts
@@ -1,13 +1,6 @@
 export * from './BalancerPoolTokenABI';
 export * from './FeederPoolABI';
 export * from './FeederPoolVaultABI';
-export * from './flatcoin/FlatcoinDelayedOrderABI';
-export * from './flatcoin/FlatcoinStableModuleABI';
-export * from './flatcoin/FlatcoinKeeperFeeABI';
-export * from './flatcoin/FlatcoinLeveragedModuleABI';
-export * from './flatcoin/FlatcoinOracleModuleABI';
-export * from './flatcoin/FlatcoinVaultABI';
-export * from './flatcoin/FlatcoinViewerABI';
 export * from './SaveABI';
 export * from './StableABI';
 export * from './StakingABI';
