+++ libs/shared/constants/src/index.ts
@@ -1,7 +1,6 @@
 export * from './abis';
 export * from './contracts';
 export * from './currency';
-export * from './flatcoin';
 export * from './metavaults';
 export * from './etherscan';
 export * from './mstable';
